+++ Example/ExampleProject/ExampleProject.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -6,7 +6,7 @@
       "location" : "https://github.com/cbepxbeo/page-viewer.git",
       "state" : {
         "branch" : "develop",
-        "revision" : "772d4f2a29df7a086411c57eb95add49b6cbb725"
+        "revision" : "f445a0cc2543b8c4c97fa46426de00a5c6d25081"
       }
     }
   ],
