+++ Sources/iOS/PageViewerPointPosition.swift
@@ -1,0 +1,12 @@
+//
+// Project: PageViewer
+// File: PageViewerPointPosition.swift
+// Created by: Egor Boyko
+// Date: 22.10.2022
+//
+// Status: #In progress | #Not decorated
+//
+
+public enum PageViewerPointPosition {
+    case bottom, top
+}
